[SMA-230] ERM Generator, Refactoring
