--- solawi-bid-frontend/src/jsMain/kotlin/org/solyton/solawi/bid/application/ui/style/button/desktop-styles.kt
@@ -1,9 +1,0 @@
-package org.solyton.solawi.bid.application.ui.style.button
-
-import org.jetbrains.compose.web.css.StyleScope
-import org.solyton.solawi.bid.application.ui.style.font.DesktopFonts
-import org.solyton.solawi.bid.application.ui.style.font.setFont
-
-val submitButtonDesktopStyle: StyleScope.()->Unit = {
-    setFont(DesktopFonts.button)
-}
