--- solawi-bid-frontend/src/jsMain/kotlin/org/solyton/solawi/bid/application/ui/style/button/mobile-styles.kt
@@ -1,11 +1,0 @@
-package org.solyton.solawi.bid.application.ui.style.button
-
-import org.jetbrains.compose.web.css.*
-import org.solyton.solawi.bid.application.ui.style.font.LargeMobileFonts
-import org.solyton.solawi.bid.application.ui.style.font.setFont
-
-val submitButtonMobileStyle: StyleScope.()->Unit = {
-    width(100.percent)
-    height(50.px)
-    setFont(LargeMobileFonts.button)
-}
