--- solawi-bid-frontend/src/jsMain/kotlin/org/solyton/solawi/bid/application/ui/style/modal/common-styles.kt
@@ -1,3 +1,0 @@
-package org.solyton.solawi.bid.application.ui.style.modal
-
-// @Style
