--- solawi-bid-frontend/src/jsMain/kotlin/org/solyton/solawi/bid/application/ui/style/modal/desktop-styles.kt
@@ -1,2 +1,0 @@
-package org.solyton.solawi.bid.application.ui.style.modal
-
