+++ solawi-bid-frontend/src/jsMain/kotlin/org/solyton/solawi/bid/module/style/modal/common-styles.kt
@@ -1,0 +1,3 @@
+package org.solyton.solawi.bid.module.style.modal
+
+// @Style
