+++ solawi-bid-frontend/src/jsMain/kotlin/org/solyton/solawi/bid/module/style/modal/mobile-styles.kt
@@ -1,0 +1,2 @@
+package org.solyton.solawi.bid.module.style.modal
+
